+++ src/main/java/io/github/karolbystrek/layers/OutputLayer.java
@@ -8,19 +8,19 @@ public class OutputLayer implements Layer {
     private final int numNodesOut;
 
     private Tensor lastInput;
-    private final double[] lastWeightedInput;
+    private final float[] lastWeightedInput;
 
-    private double[][] weightsGradient;
-    private double[] biasesGradient;
+    private float[][] weightsGradient;
+    private float[] biasesGradient;
 
-    private double[][] weights;
-    private double[] biases;
+    private float[][] weights;
+    private float[] biases;
 
     public OutputLayer(int numNodesIn, int numNodesOut) {
         this.numNodesIn = numNodesIn;
         this.numNodesOut = numNodesOut;
 
-        this.lastWeightedInput = new double[numNodesOut];
+        this.lastWeightedInput = new float[numNodesOut];
 
         initializeWeights();
         initializeBiases();
@@ -35,7 +35,7 @@ public class OutputLayer implements Layer {
         lastInput = input;
         Tensor output = new Tensor(1, 1, numNodesOut);
 
-        double maxLogit = Double.NEGATIVE_INFINITY;
+        float maxLogit = Float.NEGATIVE_INFINITY;
         for (int nodeOut = 0; nodeOut < numNodesOut; nodeOut++) {
             lastWeightedInput[nodeOut] = biases[nodeOut];
             for (int nodeIn = 0; nodeIn < numNodesIn; nodeIn++) {
@@ -47,15 +47,15 @@ public class OutputLayer implements Layer {
             }
         }
 
-        double sumExp = 0.0;
+        float sumExp = 0.0f;
         for (int nodeOut = 0; nodeOut < numNodesOut; nodeOut++) {
-            double value = Math.exp(lastWeightedInput[nodeOut] - maxLogit);
+            float value = (float) Math.exp(lastWeightedInput[nodeOut] - maxLogit);
             output.setValue(0, 0, nodeOut, value);
             sumExp += value;
         }
 
         for (int nodeOut = 0; nodeOut < numNodesOut; nodeOut++) {
-            double value = output.getValue(0, 0, nodeOut);
+            float value = output.getValue(0, 0, nodeOut);
             value /= sumExp;
             output.setValue(0, 0, nodeOut, value);
         }
@@ -65,16 +65,16 @@ public class OutputLayer implements Layer {
 
     @Override
     public Tensor backward(Tensor gradOutput) {
-        double[][][] gradOutData = gradOutput.getData();
-        double[] delta = new double[numNodesOut];
-        double[] gradInputFlat = new double[numNodesIn];
+        float[][][] gradOutData = gradOutput.getData();
+        float[] delta = new float[numNodesOut];
+        float[] gradInputFlat = new float[numNodesIn];
 
         for (int nodeOut = 0; nodeOut < numNodesOut; nodeOut++) {
             delta[nodeOut] = gradOutData[0][0][nodeOut];
             biasesGradient[nodeOut] += delta[nodeOut];
 
             for (int nodeIn = 0; nodeIn < numNodesIn; nodeIn++) {
-                double inputValue = lastInput.getValue(0, 0, nodeIn);
+                float inputValue = lastInput.getValue(0, 0, nodeIn);
                 weightsGradient[nodeOut][nodeIn] += delta[nodeOut] * inputValue;
 
                 gradInputFlat[nodeIn] += weights[nodeOut][nodeIn] * delta[nodeOut];
@@ -90,38 +90,38 @@ public class OutputLayer implements Layer {
     }
 
     @Override
-    public void updateParameters(double learningRate) {
+    public void updateParameters(float learningRate) {
         for (int nodeOut = 0; nodeOut < numNodesOut; nodeOut++) {
             biases[nodeOut] -= learningRate * biasesGradient[nodeOut];
-            biasesGradient[nodeOut] = 0.0;
+            biasesGradient[nodeOut] = 0.0f;
 
             for (int nodeIn = 0; nodeIn < numNodesIn; nodeIn++) {
                 weights[nodeOut][nodeIn] -= learningRate * weightsGradient[nodeOut][nodeIn];
-                weightsGradient[nodeOut][nodeIn] = 0.0;
+                weightsGradient[nodeOut][nodeIn] = 0.0f;
             }
         }
     }
 
     private void initializeWeights() {
-        weights = new double[numNodesOut][numNodesIn];
-        weightsGradient = new double[numNodesOut][numNodesIn];
+        weights = new float[numNodesOut][numNodesIn];
+        weightsGradient = new float[numNodesOut][numNodesIn];
 
-        double limit = Math.sqrt(1.0 / numNodesIn);
+        float limit = (float) Math.sqrt(1.0 / numNodesIn);
         for (int nodeOut = 0; nodeOut < numNodesOut; nodeOut++) {
             for (int nodeIn = 0; nodeIn < numNodesIn; nodeIn++) {
-                weights[nodeOut][nodeIn] = (Math.random() * 2 - 1) * limit;
-                weightsGradient[nodeOut][nodeIn] = 0.0;
+                weights[nodeOut][nodeIn] = (float) ((Math.random() * 2 - 1) * limit);
+                weightsGradient[nodeOut][nodeIn] = 0.0f;
             }
         }
     }
 
     private void initializeBiases() {
-        biases = new double[numNodesOut];
-        biasesGradient = new double[numNodesOut];
+        biases = new float[numNodesOut];
+        biasesGradient = new float[numNodesOut];
 
         for (int nodeOut = 0; nodeOut < numNodesOut; nodeOut++) {
-            biases[nodeOut] = 0.0;
-            biasesGradient[nodeOut] = 0.0;
+            biases[nodeOut] = 0.0f;
+            biasesGradient[nodeOut] = 0.0f;
         }
     }
 }
